Type Alias re_space_view_time_series::space_view_class::ViewType

type ViewType = TimeSeriesView;

Aliased Type

struct ViewType {
    pub axis_y: ScalarAxis,
    pub plot_legend: PlotLegend,
    pub time_ranges: VisibleTimeRanges,
}

Fields
axis_y: ScalarAxis

Configures the vertical axis of the plot.

plot_legend: PlotLegend

Configures the legend of the plot.

time_ranges: VisibleTimeRanges

Configures which range on each timeline is shown by this view (unless specified differently per entity).

If not specified, the default is to show the entire timeline. If a timeline is specified more than once, the first entry will be used.