Function re_space_view_time_series::util::next_up_f64

pub fn next_up_f64(this: f64) -> f64

Returns the least number greater than self.

Unstable feature in Rust. This is a copy of the implementation from the standard library.

Let TINY be the smallest representable positive f64. Then,

if self.is_nan(), this returns self;

if self is NEG_INFINITY, this returns MIN;

if self is -TINY, this returns -0.0;

if self is -0.0 or +0.0, this returns TINY;

if self is MAX or INFINITY, this returns INFINITY;

otherwise the unique least value greater than self is returned.

The identity x.next_up() == -(-x).next_down() holds for all non-NaN x. When x is finite x == x.next_up().next_down() also holds.

#![feature(float_next_up_down)]
// f64::EPSILON is the difference between 1.0 and the next number up.
assert_eq!(1.0f64.next_up(), 1.0 + f64::EPSILON);
// But not for most numbers.
assert!(0.1f64.next_up() < 0.1 + f64::EPSILON);
assert_eq!(9007199254740992f64.next_up(), 9007199254740994.0);